requirements.txt
django == 3.1.2
pillow
django-admin-interface
asgiref==3.3.4
bcrypt==3.2.0
cffi==1.14.6
numpy==1.21.0
pycparser==2.20
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
django-rest-framework
django-crispy-forms